(refactor) Use Guzzle instead of file_get_contents
file_get_contents returns either the response body as a string, or
false, and doesn't provide any way to get information about an error
response.

Set a user agent header so that Ops know where the requests are coming
from.

Guzzle is an OO approach to making http requests and will allow us to
inspect responses in more detail.

cURL was another approach considered, but it's a bit more bare metal and
doesn't behave like an object (e.g. you have to open a session and close
it when you're done)